+++ prg/copulas/archimedean/__init__.py
@@ -1,7 +1,13 @@
-from prg.copulas.archimedean.gumbel  import CopulaGH
-from prg.copulas.archimedean.clayton import CopulaClayton
-from prg.copulas.archimedean.frank   import CopulaFrank
-from prg.copulas.archimedean.a12     import CopulaA12
-from prg.copulas.archimedean.a14     import CopulaA14
+from prg.copulas.archimedean.gumbel    import CopulaGH
+from prg.copulas.archimedean.clayton   import CopulaClayton
+from prg.copulas.archimedean.frank     import CopulaFrank
+from prg.copulas.archimedean.a12       import CopulaA12
+from prg.copulas.archimedean.a14       import CopulaA14
+from prg.copulas.archimedean.joe       import CopulaJoe
+from prg.copulas.archimedean.survival  import SurvivalCopula, SurvivalClayton, SurvivalGH, SurvivalJoe
 
-__all__ = ['CopulaGH', 'CopulaClayton', 'CopulaFrank', 'CopulaA12', 'CopulaA14']
+__all__ = [
+    'CopulaGH', 'CopulaClayton', 'CopulaFrank', 'CopulaA12', 'CopulaA14',
+    'CopulaJoe',
+    'SurvivalCopula', 'SurvivalClayton', 'SurvivalGH', 'SurvivalJoe',
+]